+++ chain/cosmos/broadcaster.go
@@ -14,18 +14,24 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/types/tx/signing"
 	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
-	"github.com/strangelove-ventures/ibctest/broadcast"
 	"github.com/strangelove-ventures/ibctest/internal/dockerutil"
 )
 
-var _ broadcast.Broadcaster = &Broadcaster{}
+type ClientContextOpt func(clientContext client.Context) client.Context
+
+type FactoryOpt func(factory tx.Factory) tx.Factory
+
+type User interface {
+	GetKeyName() string
+	Bech32Address(bech32Prefix string) string
+}
 
 type Broadcaster struct {
 	// buf stores the output sdk.TxResponse when broadcast.Tx is invoked.
 	buf *bytes.Buffer
 	// keyrings is a mapping of keyrings which point to a temporary test directory. The contents
 	// of this directory are copied from the node container for the specific user.
-	keyrings map[broadcast.User]keyring.Keyring
+	keyrings map[User]keyring.Keyring
 
 	// chain is a reference to the CosmosChain instance which will be the target of the messages.
 	chain *CosmosChain
@@ -33,9 +39,9 @@ type Broadcaster struct {
 	t *testing.T
 
 	// factoryOptions is a slice of broadcast.FactoryOpt which enables arbitrary configuration of the tx.Factory.
-	factoryOptions []broadcast.FactoryOpt
+	factoryOptions []FactoryOpt
 	// clientContextOptions is a slice of broadcast.ClientContextOpt which enables arbitrary configuration of the client.Context.
-	clientContextOptions []broadcast.ClientContextOpt
+	clientContextOptions []ClientContextOpt
 }
 
 // NewBroadcaster returns a instance of Broadcaster which can be used with broadcast.Tx to
@@ -45,26 +51,26 @@ func NewBroadcaster(t *testing.T, chain *CosmosChain) *Broadcaster {
 		t:        t,
 		chain:    chain,
 		buf:      &bytes.Buffer{},
-		keyrings: map[broadcast.User]keyring.Keyring{},
+		keyrings: map[User]keyring.Keyring{},
 	}
 }
 
 // ConfigureFactoryOptions ensure the given configuration functions are run when calling GetFactory
 // after all default options have been applied.
-func (b *Broadcaster) ConfigureFactoryOptions(opts ...broadcast.FactoryOpt) {
+func (b *Broadcaster) ConfigureFactoryOptions(opts ...FactoryOpt) {
 	b.factoryOptions = append(b.factoryOptions, opts...)
 }
 
 // ConfigureClientContextOptions ensure the given configuration functions are run when calling GetClientContext
 // after all default options have been applied.
-func (b *Broadcaster) ConfigureClientContextOptions(opts ...broadcast.ClientContextOpt) {
+func (b *Broadcaster) ConfigureClientContextOptions(opts ...ClientContextOpt) {
 	b.clientContextOptions = append(b.clientContextOptions, opts...)
 }
 
 // GetFactory returns an instance of tx.Factory that is configured with this Broadcaster's CosmosChain
 // and the provided user. ConfigureFactoryOptions can be used to specify arbitrary options to configure the returned
 // factory.
-func (b *Broadcaster) GetFactory(ctx context.Context, user broadcast.User) (tx.Factory, error) {
+func (b *Broadcaster) GetFactory(ctx context.Context, user User) (tx.Factory, error) {
 	clientContext, err := b.GetClientContext(ctx, user)
 	if err != nil {
 		return tx.Factory{}, err
@@ -90,7 +96,7 @@ func (b *Broadcaster) GetFactory(ctx context.Context, user broadcast.User) (tx.F
 // GetClientContext returns a client context that is configured with this Broadcaster's CosmosChain and
 // the provided user. ConfigureClientContextOptions can be used to configure arbitrary options to configure the returned
 // client.Context.
-func (b *Broadcaster) GetClientContext(ctx context.Context, user broadcast.User) (client.Context, error) {
+func (b *Broadcaster) GetClientContext(ctx context.Context, user User) (client.Context, error) {
 	chain := b.chain
 	cn := chain.getFullNode()
 
@@ -118,7 +124,7 @@ func (b *Broadcaster) GetClientContext(ctx context.Context, user broadcast.User)
 }
 
 // GetTxResponseBytes returns the sdk.TxResponse bytes which returned from broadcast.Tx.
-func (b *Broadcaster) GetTxResponseBytes(ctx context.Context, user broadcast.User) ([]byte, error) {
+func (b *Broadcaster) GetTxResponseBytes(ctx context.Context, user User) ([]byte, error) {
 	if b.buf == nil || b.buf.Len() == 0 {
 		return nil, fmt.Errorf("empty buffer, transaction has not been executed yet")
 	}
@@ -136,7 +142,7 @@ func (b *Broadcaster) UnmarshalTxResponseBytes(ctx context.Context, bytes []byte
 }
 
 // defaultClientContext returns a default client context configured with the user as the sender.
-func (b *Broadcaster) defaultClientContext(fromUser broadcast.User, sdkAdd sdk.AccAddress) client.Context {
+func (b *Broadcaster) defaultClientContext(fromUser User, sdkAdd sdk.AccAddress) client.Context {
 	// initialize a clean buffer each time
 	b.buf.Reset()
 	kr := b.keyrings[fromUser]
@@ -171,4 +177,35 @@ func (b *Broadcaster) defaultTxFactory(clientCtx client.Context, accountNumber u
 		WithKeybase(clientCtx.Keyring).
 		WithChainID(clientCtx.ChainID).
 		WithSimulateAndExecute(false)
+}
+
+// BroadcastTx uses the provided Broadcaster to broadcast all the provided messages which will be signed
+// by the User provided. The sdk.TxResponse and an error are returned.
+func BroadcastTx(ctx context.Context, broadcaster *Broadcaster, broadcastingUser User, msgs ...sdk.Msg) (sdk.TxResponse, error) {
+	for _, msg := range msgs {
+		if err := msg.ValidateBasic(); err != nil {
+			return sdk.TxResponse{}, err
+		}
+	}
+
+	f, err := broadcaster.GetFactory(ctx, broadcastingUser)
+	if err != nil {
+		return sdk.TxResponse{}, err
+	}
+
+	cc, err := broadcaster.GetClientContext(ctx, broadcastingUser)
+	if err != nil {
+		return sdk.TxResponse{}, err
+	}
+
+	if err := tx.BroadcastTx(cc, f, msgs...); err != nil {
+		return sdk.TxResponse{}, err
+	}
+
+	txBytes, err := broadcaster.GetTxResponseBytes(ctx, broadcastingUser)
+	if err != nil {
+		return sdk.TxResponse{}, err
+	}
+
+	return broadcaster.UnmarshalTxResponseBytes(ctx, txBytes)
 }
